+++ server/src-lib/Hasura/Backends/Postgres/Translate/Select.hs
@@ -16,7 +16,7 @@ import qualified Data.List.NonEmpty                         as NE
 import qualified Data.Text                                  as T
 import qualified Database.PG.Query                          as Q
 
-import           Control.Lens                               hiding (op)
+import           Control.Lens                               ((^?))
 import           Control.Monad.Writer.Strict
 import           Data.Text.Extended
 
@@ -308,6 +308,7 @@ mkAggregateOrderByExtractorAndFields annAggOrderBy =
   where
     alias = Just $ mkAggregateOrderByAlias annAggOrderBy
 
+-- | Generate alias for order by extractors
 mkAnnOrderByAlias
   :: Identifier -> FieldName -> SimilarArrayFields -> AnnotatedOrderByElement ('Postgres pgKind) v -> S.Alias
 mkAnnOrderByAlias pfx parAls similarFields = \case
@@ -337,13 +338,18 @@ mkAnnOrderByAlias pfx parAls similarFields = \case
             obAls = cfPfx <> Identifier "." <> toIdentifier (mkAggregateOrderByAlias aggOrderBy)
         in S.Alias obAls
 
-processDistinctOnColumns
+applyDistinctOnAtBase
+  :: NE.NonEmpty PGCol -> S.DistinctExpr
+applyDistinctOnAtBase =
+  S.DistinctOn . map (S.SEIdentifier . toIdentifier) . toList
+
+applyDistinctOnAtNode
   :: Identifier
   -> NE.NonEmpty PGCol
   -> ( S.DistinctExpr
      , [(S.Alias, S.SQLExp)] -- additional column extractors
      )
-processDistinctOnColumns pfx neCols = (distOnExp, colExtrs)
+applyDistinctOnAtNode pfx neCols = (distOnExp, colExtrs)
   where
     cols = toList neCols
     distOnExp = S.DistinctOn $ map (S.SEIdentifier . toIdentifier . mkQColAls) cols
@@ -496,8 +502,14 @@ processAnnSimpleSelect
        )
 processAnnSimpleSelect sourcePrefixes fieldAlias permLimitSubQuery annSimpleSel = do
   (selectSource, orderByAndDistinctExtrs, _) <-
-    processSelectParams sourcePrefixes fieldAlias similarArrayFields tableFrom
-    permLimitSubQuery tablePermissions tableArgs
+    processSelectParams
+      sourcePrefixes
+      fieldAlias
+      similarArrayFields
+      tableFrom
+      permLimitSubQuery
+      tablePermissions
+      tableArgs
   annFieldsExtr <- processAnnFields (_pfThis sourcePrefixes) fieldAlias similarArrayFields annSelFields
   let allExtractors = HM.fromList $ annFieldsExtr : orderByAndDistinctExtrs
   pure (selectSource, allExtractors)
@@ -522,8 +534,14 @@ processAnnAggregateSelect
        )
 processAnnAggregateSelect sourcePrefixes fieldAlias annAggSel = do
   (selectSource, orderByAndDistinctExtrs, _) <-
-    processSelectParams sourcePrefixes fieldAlias similarArrayFields tableFrom
-    permLimitSubQuery tablePermissions tableArgs
+    processSelectParams
+      sourcePrefixes
+      fieldAlias
+      similarArrayFields
+      tableFrom
+      permLimitSubQuery
+      tablePermissions
+      tableArgs
   let thisSourcePrefix = _pfThis sourcePrefixes
   processedFields <- forM aggSelFields $ \(fieldName, field) ->
     (fieldName,) <$>
@@ -535,7 +553,7 @@ processAnnAggregateSelect sourcePrefixes fieldAlias annAggSel = do
       TAFNodes _ annFields -> do
         annFieldExtr <- processAnnFields thisSourcePrefix fieldName similarArrayFields annFields
         pure ( [annFieldExtr]
-             , withJsonAggExtr permLimitSubQuery (_ssOrderBy selectSource) $
+             , withJsonAggExtr permLimitSubQuery (orderByForJsonAgg selectSource) $
                S.Alias $ toIdentifier fieldName
              )
       TAFExp e ->
@@ -608,7 +626,7 @@ processArrayRelation sourcePrefixes fieldAlias relAlias arrSel =
       (source, nodeExtractors) <-
         processAnnSimpleSelect sourcePrefixes fieldAlias permLimitSubQuery sel
       let topExtr = asJsonAggExtr JASMultipleRows (S.toAlias fieldAlias)
-                    permLimitSubQuery $ _ssOrderBy source
+                    permLimitSubQuery $ orderByForJsonAgg source
       pure ( ArrayRelationSource relAlias colMapping source
            , topExtr
            , nodeExtractors
@@ -633,6 +651,24 @@ processArrayRelation sourcePrefixes fieldAlias relAlias arrSel =
            , ()
            )
 
+{- Note [Optimizing queries using limit/offset]
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+Refer to the issue https://github.com/hasura/graphql-engine/issues/5745
+
+Before this change, limit/offset/distinct_on is applied at outer selection
+node along with order by clause. This greatly reduces query performance if
+our base selection table contains many rows and relationships are selected
+which joins remote tables. We need to optimize application of limit wrt to
+order by input.
+
+If "Order by" is not present:
+  Apply limit/offset/distinct on at the base table selection
+Else if "Order by" contains only columns:
+  Apply limit/offset/distinct_on at the base table selection along with order by
+Otherwise:
+  Apply limit/offset/distinct_on at the node selection along with order by
+-}
+
 processSelectParams
   :: forall pgKind m
    . ( MonadReader Bool m
@@ -652,25 +688,23 @@ processSelectParams
        )
 processSelectParams sourcePrefixes fieldAlias similarArrFields selectFrom
                     permLimitSubQ tablePermissions tableArgs = do
-  maybeOrderBy <- mapM
-                  (processOrderByItems thisSourcePrefix fieldAlias similarArrFields)
-                  orderByM
+  (additionalExtrs, selectSorting, cursorExp) <-
+    processOrderByItems thisSourcePrefix fieldAlias similarArrFields distM orderByM
   let fromItem = selectFromToFromItem (_pfBase sourcePrefixes) selectFrom
-      (maybeDistinct, distinctExtrs) =
-        maybe (Nothing, []) (first Just) $ processDistinctOnColumns thisSourcePrefix <$> distM
       finalWhere = toSQLBoolExp (selectFromToQual selectFrom) $
                    maybe permFilter (andAnnBoolExps permFilter) whereM
-      selectSource = SelectSource thisSourcePrefix fromItem maybeDistinct finalWhere
-                     ((^. _2) <$> maybeOrderBy) finalLimit offsetM
-      orderByExtrs = maybe [] (^. _1) maybeOrderBy
+      sortingAndSlicing = SortingAndSlicing selectSorting selectSlicing
+      selectSource = SelectSource thisSourcePrefix fromItem finalWhere
+                     sortingAndSlicing
   pure ( selectSource
-       , orderByExtrs <> distinctExtrs
-       , (^. _3) <$> maybeOrderBy
+       , additionalExtrs
+       , cursorExp
        )
   where
     thisSourcePrefix = _pfThis sourcePrefixes
     SelectArgs whereM orderByM inpLimitM offsetM distM = tableArgs
     TablePerm permFilter permLimit = tablePermissions
+    selectSlicing = SelectSlicing finalLimit offsetM
     finalLimit =
     -- if sub query is required, then only use input limit
     --    because permission limit is being applied in subquery
@@ -683,7 +717,7 @@ processSelectParams sourcePrefixes fieldAlias similarArrFields selectFrom
       case (inpLimitM, permLimit) of
         (inpLim, Nothing)     -> inpLim
         (Nothing, permLim)    -> permLim
-        (Just inp, Just perm) -> Just $ if inp < perm then inp else perm
+        (Just inp, Just perm) -> Just if inp < perm then inp else perm
 
 processOrderByItems
   :: forall pgKind m
@@ -694,17 +728,20 @@ processOrderByItems
   => Identifier
   -> FieldName
   -> SimilarArrayFields
-  -> NE.NonEmpty (AnnotatedOrderByItem ('Postgres pgKind))
+  -> Maybe (NE.NonEmpty PGCol)
+  -> Maybe (NE.NonEmpty (AnnotatedOrderByItem ('Postgres pgKind)))
   -> m ( [(S.Alias, S.SQLExp)] -- Order by Extractors
-       , S.OrderByExp
-       , S.SQLExp -- The cursor expression
+       , SelectSorting
+       , Maybe S.SQLExp -- The cursor expression
        )
-processOrderByItems sourcePrefix' fieldAlias' similarArrayFields orderByItems = do
-  orderByItemExps <- forM orderByItems processAnnOrderByItem
-  let orderByExp = S.OrderByExp $ toOrderByExp <$> orderByItemExps
-      orderByExtractors = concat $ toList $ map snd . toList <$> orderByItemExps
-      cursor = mkCursorExp $ toList orderByItemExps
-  pure (orderByExtractors, orderByExp, cursor)
+processOrderByItems sourcePrefix' fieldAlias' similarArrayFields distOnCols = \case
+  Nothing -> pure ([], NoSorting $ applyDistinctOnAtBase <$> distOnCols, Nothing)
+  Just orderByItems -> do
+    orderByItemExps <- forM orderByItems processAnnOrderByItem
+    let (sorting, distinctOnExtractors) = generateSorting orderByItemExps
+        orderByExtractors = concat $ toList $ map snd . toList <$> orderByItemExps
+        cursor = mkCursorExp $ toList orderByItemExps
+    pure (orderByExtractors <> distinctOnExtractors, sorting, Just cursor)
   where
     processAnnOrderByItem
       :: AnnotatedOrderByItem ('Postgres pgKind)
@@ -714,7 +751,7 @@ processOrderByItems sourcePrefix' fieldAlias' similarArrayFields orderByItems = 
       processAnnotatedOrderByElement sourcePrefix' fieldAlias' ordByCol
 
     processAnnotatedOrderByElement
-      :: Identifier -> FieldName -> AnnotatedOrderByElement ('Postgres pgKind) S.SQLExp -> m (S.Alias, S.SQLExp)
+      :: Identifier -> FieldName -> AnnotatedOrderByElement ('Postgres pgKind) S.SQLExp -> m (S.Alias, (SQLExpression ('Postgres pgKind)))
     processAnnotatedOrderByElement sourcePrefix fieldAlias annObCol = do
       let ordByAlias = mkAnnOrderByAlias sourcePrefix fieldAlias similarArrayFields annObCol
       (ordByAlias, ) <$> case annObCol of
@@ -744,9 +781,9 @@ processOrderByItems sourcePrefix' fieldAlias' similarArrayFields orderByItems = 
               relAlias = mkArrayRelationAlias fieldAlias similarArrayFields fieldName
               (topExtractor, fields) = mkAggregateOrderByExtractorAndFields aggOrderBy
               selectSource = SelectSource relSourcePrefix
-                             (S.FISimple relTable Nothing) Nothing
+                             (S.FISimple relTable Nothing)
                              (toSQLBoolExp (S.QualTable relTable) relFilter)
-                             Nothing Nothing Nothing
+                             noSortingAndSlicing
               relSource = ArrayRelationSource relAlias colMapping selectSource
           pure ( relSource
                , topExtractor
@@ -767,9 +804,9 @@ processOrderByItems sourcePrefix' fieldAlias' similarArrayFields orderByItems = 
                   fromItem = selectFromToFromItem sourcePrefix $
                              FromFunction _cfobFunction _cfobFunctionArgsExp Nothing
                   functionQual = S.QualifiedIdentifier (functionToIdentifier _cfobFunction) Nothing
-                  selectSource = SelectSource computedFieldSourcePrefix fromItem Nothing
+                  selectSource = SelectSource computedFieldSourcePrefix fromItem
                                  (toSQLBoolExp functionQual tableFilter)
-                                 Nothing Nothing Nothing
+                                 noSortingAndSlicing
                   source = ComputedFieldTableSetSource fieldName selectSource
               pure ( source
                    , topExtractor
@@ -777,12 +814,54 @@ processOrderByItems sourcePrefix' fieldAlias' similarArrayFields orderByItems = 
                    , S.mkQIdenExp computedFieldSourcePrefix (mkAggregateOrderByAlias aggOrderBy)
                    )
 
-    toOrderByExp
-      :: OrderByItemG ('Postgres pgKind) (AnnotatedOrderByElement ('Postgres pgKind) (SQLExpression ('Postgres pgKind)), (S.Alias, SQLExpression ('Postgres pgKind)))
-      -> S.OrderByItem
-    toOrderByExp orderByItemExp =
-      let OrderByItemG obTyM expAlias obNullsM = fst . snd <$> orderByItemExp
-      in S.OrderByItem (S.SEIdentifier $ toIdentifier expAlias) obTyM obNullsM
+    -- | To sort queries, the 'ORDER BY' clause is used. The order by clause is applied either at base table selection
+    -- or outer node selection based on whether the list of order by expression contains non-column order by item.
+    -- For more details see Note [Optimizing queries using limit/offset].
+    generateSorting
+      :: NE.NonEmpty (OrderByItemG ('Postgres pgKind) (AnnotatedOrderByElement ('Postgres pgKind) (SQLExpression ('Postgres pgKind)), (S.Alias, SQLExpression ('Postgres pgKind))))
+      -> ( SelectSorting
+         , [(S.Alias, SQLExpression ('Postgres pgKind))] -- 'distinct on' column extractors
+         )
+    generateSorting orderByExps@(firstOrderBy NE.:| restOrderBys) =
+      case fst $ obiColumn firstOrderBy of
+        AOCColumn columnInfo ->
+          -- If rest order by expressions are all columns then apply order by clause at base selection.
+          if all (isJust . getColumnOrderBy . obiColumn) restOrderBys then
+            -- Collect column order by expressions from the rest.
+            let restColumnOrderBys = mapMaybe (sequenceA . (getColumnOrderBy <$>)) restOrderBys
+                firstColumnOrderBy = firstOrderBy{obiColumn = columnInfo}
+            in sortAtNodeAndBase $ firstColumnOrderBy NE.:| restColumnOrderBys
+
+          -- Else rest order by expressions contain atleast one non-column order by.
+          -- So, apply order by clause at node selection.
+          else sortOnlyAtNode
+        _                    ->
+          -- First order by expression is non-column. So, apply order by clause at node selection.
+          sortOnlyAtNode
+
+      where
+        getColumnOrderBy = (^? _AOCColumn) . fst
+
+        (nodeOrderBy, nodeDistinctOn, nodeDistinctOnExtractors) =
+          let toOrderByExp orderByItemExp =
+                let OrderByItemG obTyM expAlias obNullsM = fst . snd <$> orderByItemExp
+                in S.OrderByItem (S.SEIdentifier $ toIdentifier expAlias) obTyM obNullsM
+              orderByExp = S.OrderByExp $ toOrderByExp <$> orderByExps
+              (maybeDistOn, distOnExtrs) = NE.unzip $ applyDistinctOnAtNode sourcePrefix' <$> distOnCols
+          in (orderByExp, maybeDistOn, fromMaybe [] distOnExtrs)
+
+        sortOnlyAtNode =
+          (Sorting $ ASorting (nodeOrderBy, nodeDistinctOn) Nothing, nodeDistinctOnExtractors)
+
+        sortAtNodeAndBase baseColumnOrderBys =
+          let mkBaseOrderByItem (OrderByItemG orderByType columnInfo nullsOrder) =
+                S.OrderByItem
+                (S.SEIdentifier $ toIdentifier $ pgiColumn columnInfo)
+                orderByType nullsOrder
+              baseOrderByExp = S.OrderByExp $ mkBaseOrderByItem <$> baseColumnOrderBys
+              baseDistOnExp = applyDistinctOnAtBase <$> distOnCols
+              sorting = Sorting $ ASorting (nodeOrderBy, nodeDistinctOn) $ Just (baseOrderByExp, baseDistOnExp)
+          in (sorting, nodeDistinctOnExtractors)
 
     mkCursorExp
       :: [OrderByItemG ('Postgres pgKind) (AnnotatedOrderByElement ('Postgres pgKind) (SQLExpression ('Postgres pgKind)), (S.Alias, SQLExpression ('Postgres pgKind)))]
@@ -960,7 +1039,7 @@ processAnnFields sourcePrefix fieldAlias similarArrFields annFields = do
           fieldName PLSQNotRequired sel
         let computedFieldTableSetSource = ComputedFieldTableSetSource fieldName selectSource
             extractor = asJsonAggExtr selectTy (S.toAlias fieldName) PLSQNotRequired $
-                        _ssOrderBy selectSource
+                        orderByForJsonAgg selectSource
         pure ( computedFieldTableSetSource
              , extractor
              , nodeExtractors
@@ -1038,21 +1117,27 @@ generateSQLSelect joinCondition selectSource selectNode =
   S.mkSelect
   { S.selExtr = [S.Extractor e $ Just a | (a, e) <- HM.toList extractors]
   , S.selFrom = Just $ S.FromExp [joinedFrom]
-  , S.selOrderBy = maybeOrderby
-  , S.selLimit = S.LimitExp . S.intToSQLExp <$> maybeLimit
-  , S.selOffset = S.OffsetExp . S.int64ToSQLExp <$> maybeOffset
-  , S.selDistinct = maybeDistinct
+  , S.selOrderBy = nodeOrderBy
+  , S.selLimit = S.LimitExp . S.intToSQLExp <$> _ssLimit nodeSlicing
+  , S.selOffset = S.OffsetExp . S.int64ToSQLExp <$> _ssOffset nodeSlicing
+  , S.selDistinct = nodeDistinctOn
   }
   where
-    SelectSource sourcePrefix fromItem maybeDistinct whereExp
-      maybeOrderby maybeLimit maybeOffset = selectSource
+    SelectSource sourcePrefix fromItem whereExp sortAndSlice = selectSource
     SelectNode extractors joinTree = selectNode
     JoinTree objectRelations arrayRelations arrayConnections computedFields = joinTree
+    ApplySortingAndSlicing (baseOrderBy, baseSlicing, baseDistinctOn)
+      (nodeOrderBy, nodeSlicing, nodeDistinctOn) = applySortingAndSlicing sortAndSlice
+
     -- this is the table which is aliased as "sourcePrefix.base"
     baseSelect = S.mkSelect
       { S.selExtr  = [S.Extractor (S.SEStar Nothing) Nothing]
       , S.selFrom  = Just $ S.FromExp [fromItem]
       , S.selWhere = Just $ injectJoinCond joinCondition whereExp
+      , S.selOrderBy = baseOrderBy
+      , S.selLimit = S.LimitExp . S.intToSQLExp <$> _ssLimit baseSlicing
+      , S.selOffset = S.OffsetExp . S.int64ToSQLExp <$> _ssOffset baseSlicing
+      , S.selDistinct = baseDistinctOn
       }
     baseSelectAlias = S.Alias $ mkBaseTableAlias sourcePrefix
     baseFromItem = S.mkSelFromItem baseSelect baseSelectAlias
@@ -1159,7 +1244,7 @@ mkSQLSelect jsonAggSelect annSel =
         processAnnSimpleSelect sourcePrefixes rootFldName permLimitSubQuery annSel
       selectNode = SelectNode nodeExtractors joinTree
       topExtractor = asJsonAggExtr jsonAggSelect rootFldAls permLimitSubQuery
-                     $ _ssOrderBy selectSource
+                     $ orderByForJsonAgg selectSource
       arrayNode = MultiRowSelectNode [topExtractor] selectNode
   in prefixNumToAliases $
      generateSQLSelectFromArrayNode selectSource arrayNode $ S.BELit True
@@ -1260,8 +1345,14 @@ processConnectionSelect
        )
 processConnectionSelect sourcePrefixes fieldAlias relAlias colMapping connectionSelect = do
   (selectSource, orderByAndDistinctExtrs, maybeOrderByCursor) <-
-    processSelectParams sourcePrefixes fieldAlias similarArrayFields selectFrom
-    permLimitSubQuery tablePermissions tableArgs
+    processSelectParams
+      sourcePrefixes
+      fieldAlias
+      similarArrayFields
+      selectFrom
+      permLimitSubQuery
+      tablePermissions
+      tableArgs
 
   let mkCursorExtractor = (S.Alias cursorIdentifier,) . (`S.SETyAnn` S.textTypeAnn)
       cursorExtractors = case maybeOrderByCursor of
@@ -1270,8 +1361,7 @@ processConnectionSelect sourcePrefixes fieldAlias relAlias colMapping connection
           -- Extract primary key columns from base select along with cursor expression.
           -- Those columns are required to perform connection split via a WHERE clause.
           mkCursorExtractor primaryKeyColumnsObjectExp : primaryKeyColumnExtractors
-      orderByExp = _ssOrderBy selectSource
-  (topExtractorExp, exps) <- flip runStateT [] $ processFields orderByExp
+  (topExtractorExp, exps) <- flip runStateT [] $ processFields selectSource
   let topExtractor = S.Extractor topExtractorExp $ Just $ S.Alias fieldIdentifier
       allExtractors = HM.fromList $ cursorExtractors <> exps <> orderByAndDistinctExtrs
       arrayConnectionSource = ArrayConnectionSource relAlias colMapping
@@ -1347,8 +1437,8 @@ processConnectionSelect sourcePrefixes fieldAlias relAlias colMapping connection
          , MonadWriter JoinTree n
          , MonadState [(S.Alias, S.SQLExp)] n
          )
-      => Maybe S.OrderByExp -> n S.SQLExp
-    processFields orderByExp =
+      => SelectSource -> n S.SQLExp
+    processFields selectSource =
       fmap (S.applyJsonBuildObj . concat) $
       forM fields $
         \(FieldName fieldText, field) -> (S.SELit fieldText:) . pure <$>
@@ -1356,7 +1446,7 @@ processConnectionSelect sourcePrefixes fieldAlias relAlias colMapping connection
           ConnectionTypename t              -> pure $ withForceAggregation S.textTypeAnn $ S.SELit t
           ConnectionPageInfo pageInfoFields -> pure $ processPageInfoFields pageInfoFields
           ConnectionEdges edges ->
-            fmap (flip mkSimpleJsonAgg orderByExp . S.applyJsonBuildObj . concat) $ forM edges $
+            fmap (flip mkSimpleJsonAgg (orderByForJsonAgg selectSource) . S.applyJsonBuildObj . concat) $ forM edges $
             \(FieldName edgeText, edge) -> (S.SELit edgeText:) . pure <$>
             case edge of
               EdgeTypename t -> pure $ S.SELit t